+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added video frame to LAN.
 - Added link(s) to active LAN(s) on front page.
 - Added links to both user profile and LAN seating in the arrivals list.
+- Added LAN listing for competitions.
 
 ### Changed
 - Made footer prettier.